requirements.txt
# Core Bot
python-telegram-bot[webhooks]==21.5
httpx==0.27.2

# Database
motor==3.5.1
pymongo==4.8.0

# Encryption
cryptography==43.0.1

# Scheduler
APScheduler==3.10.4

# Config
python-dotenv==1.0.1
